+++ netdef/Engines/webadmin/Views.py
@@ -37,6 +37,7 @@ class Views():
         activate_views["StatisticsModel"] = 1
         activate_views["Tools"] = 1
         activate_views["SecurityWebadminView"] = 1
+        activate_views["SecurityCertificatesView"] = 1
         activate_views.update(self.shared.config.get_dict("webadmin_views"))
 
         for name, activate in activate_views.items():