+++ CHANGELOG.rst
@@ -5,7 +5,7 @@
 
 - OPCUAClientController: Improved configuration
 - OPCUAServerController: Added legacy support for basic128rsa15 and basic256.
-- Webadmin: Added SecurityWebadminView
+- Webadmin: Added SecurityWebadminView and SecurityCertificatesView
 
 **Bug fixes**